re_grpc_proto::build::bazel::remote::execution

Module v2

Modules

action_cache_client

Generated client implementations.

action_cache_server

Generated server implementations.

batch_read_blobs_response

Nested message and enum types in BatchReadBlobsResponse.

batch_update_blobs_request

Nested message and enum types in BatchUpdateBlobsRequest.

batch_update_blobs_response

Nested message and enum types in BatchUpdateBlobsResponse.

capabilities_client

Generated client implementations.

capabilities_server

Generated server implementations.

command

Nested message and enum types in Command.

compressor

Nested message and enum types in Compressor.

content_addressable_storage_client

Generated client implementations.

content_addressable_storage_server

Generated server implementations.

digest_function

Nested message and enum types in DigestFunction.

execution_client

Generated client implementations.

execution_server

Generated server implementations.

execution_stage

Nested message and enum types in ExecutionStage.

platform

Nested message and enum types in Platform.

priority_capabilities

Nested message and enum types in PriorityCapabilities.

symlink_absolute_path_strategy

Nested message and enum types in SymlinkAbsolutePathStrategy.

Structs

Action

An Action captures all the information about an execution which is required to reproduce it.

ActionCacheUpdateCapabilities

Describes the server/instance capabilities for updating the action cache.

ActionResult

An ActionResult represents the result of an [Action][build.bazel.remote.execution.v2.Action] being run.

BatchReadBlobsRequest

A request message for [ContentAddressableStorage.BatchReadBlobs][build.bazel.remote.execution.v2.ContentAddressableStorage.BatchReadBlobs].

BatchReadBlobsResponse

A response message for [ContentAddressableStorage.BatchReadBlobs][build.bazel.remote.execution.v2.ContentAddressableStorage.BatchReadBlobs].

BatchUpdateBlobsRequest

A request message for [ContentAddressableStorage.BatchUpdateBlobs][build.bazel.remote.execution.v2.ContentAddressableStorage.BatchUpdateBlobs].

BatchUpdateBlobsResponse

A response message for [ContentAddressableStorage.BatchUpdateBlobs][build.bazel.remote.execution.v2.ContentAddressableStorage.BatchUpdateBlobs].

CacheCapabilities

Capabilities of the remote cache system.

Command

A Command is the actual command executed by a worker running an [Action][build.bazel.remote.execution.v2.Action] and specifications of its environment.

Compressor

Compression formats which may be supported.

Digest

A content digest. A digest for a given blob consists of the size of the blob and its hash. The hash algorithm to use is defined by the server.

DigestFunction

The digest function used for converting values into keys for CAS and Action Cache.

Directory

A Directory represents a directory node in a file tree, containing zero or more children [FileNodes][build.bazel.remote.execution.v2.FileNode], [DirectoryNodes][build.bazel.remote.execution.v2.DirectoryNode] and [SymlinkNodes][build.bazel.remote.execution.v2.SymlinkNode]. Each Node contains its name in the directory, either the digest of its content (either a file blob or a Directory proto) or a symlink target, as well as possibly some metadata about the file or directory.

DirectoryNode

A DirectoryNode represents a child of a [Directory][build.bazel.remote.execution.v2.Directory] which is itself a Directory and its associated metadata.

ExecuteOperationMetadata

Metadata about an ongoing [execution][build.bazel.remote.execution.v2.Execution.Execute], which will be contained in the [metadata field][google.longrunning.Operation.response] of the [Operation][google.longrunning.Operation].

ExecuteRequest

A request message for [Execution.Execute][build.bazel.remote.execution.v2.Execution.Execute].

ExecuteResponse

The response message for [Execution.Execute][build.bazel.remote.execution.v2.Execution.Execute], which will be contained in the [response field][google.longrunning.Operation.response] of the [Operation][google.longrunning.Operation].

ExecutedActionMetadata

ExecutedActionMetadata contains details about a completed execution.

ExecutionCapabilities

Capabilities of the remote execution system.

ExecutionPolicy

An ExecutionPolicy can be used to control the scheduling of the action.

ExecutionStage

The current stage of action execution.

FileNode

A FileNode represents a single file and associated metadata.

FindMissingBlobsRequest

A request message for [ContentAddressableStorage.FindMissingBlobs][build.bazel.remote.execution.v2.ContentAddressableStorage.FindMissingBlobs].

FindMissingBlobsResponse

A response message for [ContentAddressableStorage.FindMissingBlobs][build.bazel.remote.execution.v2.ContentAddressableStorage.FindMissingBlobs].

GetActionResultRequest

A request message for [ActionCache.GetActionResult][build.bazel.remote.execution.v2.ActionCache.GetActionResult].

GetCapabilitiesRequest

A request message for [Capabilities.GetCapabilities][build.bazel.remote.execution.v2.Capabilities.GetCapabilities].

GetTreeRequest

A request message for [ContentAddressableStorage.GetTree][build.bazel.remote.execution.v2.ContentAddressableStorage.GetTree].

GetTreeResponse

A response message for [ContentAddressableStorage.GetTree][build.bazel.remote.execution.v2.ContentAddressableStorage.GetTree].

LogFile

A LogFile is a log stored in the CAS.

NodeProperties

Node properties for [FileNodes][build.bazel.remote.execution.v2.FileNode], [DirectoryNodes][build.bazel.remote.execution.v2.DirectoryNode], and [SymlinkNodes][build.bazel.remote.execution.v2.SymlinkNode]. The server is responsible for specifying the properties that it accepts.

NodeProperty

A single property for [FileNodes][build.bazel.remote.execution.v2.FileNode], [DirectoryNodes][build.bazel.remote.execution.v2.DirectoryNode], and [SymlinkNodes][build.bazel.remote.execution.v2.SymlinkNode]. The server is responsible for specifying the property names that it accepts. If permitted by the server, the same name may occur multiple times.

OutputDirectory

An OutputDirectory is the output in an ActionResult corresponding to a directory’s full contents rather than a single file.

OutputFile

An OutputFile is similar to a [FileNode][build.bazel.remote.execution.v2.FileNode], but it is used as an output in an ActionResult. It allows a full file path rather than only a name.

OutputSymlink

An OutputSymlink is similar to a [Symlink][build.bazel.remote.execution.v2.SymlinkNode], but it is used as an output in an ActionResult.

Platform

A Platform is a set of requirements, such as hardware, operating system, or compiler toolchain, for an [Action][build.bazel.remote.execution.v2.Action]’s execution environment. A Platform is represented as a series of key-value pairs representing the properties that are required of the platform.

PriorityCapabilities

Allowed values for priority in [ResultsCachePolicy][build.bazel.remoteexecution.v2.ResultsCachePolicy] and [ExecutionPolicy][build.bazel.remoteexecution.v2.ResultsCachePolicy] Used for querying both cache and execution valid priority ranges.

RequestMetadata

An optional Metadata to attach to any RPC request to tell the server about an external context of the request. The server may use this for logging or other purposes. To use it, the client attaches the header to the call using the canonical proto serialization:

ResultsCachePolicy

A ResultsCachePolicy is used for fine-grained control over how action outputs are stored in the CAS and Action Cache.

ServerCapabilities

A response message for [Capabilities.GetCapabilities][build.bazel.remote.execution.v2.Capabilities.GetCapabilities].

SymlinkAbsolutePathStrategy

Describes how the server treats absolute symlink targets.

SymlinkNode

A SymlinkNode represents a symbolic link.

ToolDetails

Details for the tool used to call the API.

Tree

A Tree contains all the [Directory][build.bazel.remote.execution.v2.Directory] protos in a single directory Merkle tree, compressed into one message.

UpdateActionResultRequest

A request message for [ActionCache.UpdateActionResult][build.bazel.remote.execution.v2.ActionCache.UpdateActionResult].

WaitExecutionRequest

A request message for [WaitExecution][build.bazel.remote.execution.v2.Execution.WaitExecution].