re_grpc_proto::build::bazel::remote::execution::v2::action_cache_server

Trait ActionCache

pub trait ActionCache:
    Send
    + Sync
    + 'static {
    // Required methods
    fn get_action_result<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetActionResultRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ActionResult>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn update_action_result<'life0, 'async_trait>(
        &'life0 self,
        request: Request<UpdateActionResultRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ActionResult>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with ActionCacheServer.

Required Methods

fn get_action_result<'life0, 'async_trait>( &'life0 self, request: Request<GetActionResultRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ActionResult>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieve a cached execution result.

Implementations SHOULD ensure that any blobs referenced from the [ContentAddressableStorage][build.bazel.remote.execution.v2.ContentAddressableStorage] are available at the time of returning the [ActionResult][build.bazel.remote.execution.v2.ActionResult] and will be for some period of time afterwards. The lifetimes of the referenced blobs SHOULD be increased if necessary and applicable.

Errors:

NOT_FOUND: The requested ActionResult is not in the cache.

fn update_action_result<'life0, 'async_trait>( &'life0 self, request: Request<UpdateActionResultRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ActionResult>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Upload a new execution result.

In order to allow the server to perform access control based on the type of action, and to assist with client debugging, the client MUST first upload the [Action][build.bazel.remote.execution.v2.Execution] that produced the result, along with its [Command][build.bazel.remote.execution.v2.Command], into the ContentAddressableStorage.

Server implementations MAY modify the UpdateActionResultRequest.action_result and return an equivalent value.

Errors:

INVALID_ARGUMENT: One or more arguments are invalid.

FAILED_PRECONDITION: One or more errors occurred in updating the action result, such as a missing command or action.

RESOURCE_EXHAUSTED: There is insufficient storage space to add the entry to the cache.