re_grpc_proto::build::bazel::remote::execution::v2::capabilities_server

Trait Capabilities

pub trait Capabilities:
    Send
    + Sync
    + 'static {
    // Required method
    fn get_capabilities<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetCapabilitiesRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ServerCapabilities>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with CapabilitiesServer.

Required Methods

fn get_capabilities<'life0, 'async_trait>( &'life0 self, request: Request<GetCapabilitiesRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ServerCapabilities>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

GetCapabilities returns the server capabilities configuration of the remote endpoint. Only the capabilities of the services supported by the endpoint will be returned:

Execution + CAS + Action Cache endpoints should return both CacheCapabilities and ExecutionCapabilities.

Execution only endpoints should return ExecutionCapabilities.

CAS + Action Cache only endpoints should return CacheCapabilities.

There are no method-specific errors.