re_grpc_proto::build::bazel::remote::execution::v2::content_addressable_storage_client

Struct ContentAddressableStorageClient

pub struct ContentAddressableStorageClient<T> { /* private fields */ }

The CAS (content-addressable storage) is used to store the inputs to and outputs from the execution service. Each piece of content is addressed by the digest of its binary data.

Most of the binary data stored in the CAS is opaque to the execution engine, and is only used as a communication medium. In order to build an [Action][build.bazel.remote.execution.v2.Action], however, the client will need to also upload the [Command][build.bazel.remote.execution.v2.Command] and input root [Directory][build.bazel.remote.execution.v2.Directory] for the Action. The Command and Directory messages must be marshalled to wire format and then uploaded under the hash as with any other piece of content. In practice, the input root directory is likely to refer to other Directories in its hierarchy, which must also each be uploaded on their own.

For small file uploads the client should group them together and call [BatchUpdateBlobs][build.bazel.remote.execution.v2.ContentAddressableStorage.BatchUpdateBlobs].

For large uploads, the client must use the [Write method][google.bytestream.ByteStream.Write] of the ByteStream API.

For uncompressed data, The WriteRequest.resource_name is of the following form: {instance_name}/uploads/{uuid}/blobs/{hash}/{size}{/optional_metadata}

Where:

instance_name is an identifier, possibly containing multiple path segments, used to distinguish between the various instances on the server, in a manner defined by the server. If it is the empty path, the leading slash is omitted, so that the resource_name becomes uploads/{uuid}/blobs/{hash}/{size}{/optional_metadata}. To simplify parsing, a path segment cannot equal any of the following keywords: blobs, uploads, actions, actionResults, operations, capabilities or compressed-blobs.

uuid is a version 4 UUID generated by the client, used to avoid collisions between concurrent uploads of the same data. Clients MAY reuse the same uuid for uploading different blobs.

hash and size refer to the [Digest][build.bazel.remote.execution.v2.Digest] of the data being uploaded.

optional_metadata is implementation specific data, which clients MAY omit. Servers MAY ignore this metadata.

Data can alternatively be uploaded in compressed form, with the following WriteRequest.resource_name form: {instance_name}/uploads/{uuid}/compressed-blobs/{compressor}/{uncompressed_hash}/{uncompressed_size}{/optional_metadata}

Where:

instance_name, uuid and optional_metadata are defined as above.

compressor is a lowercase string form of a Compressor.Value enum other than identity, which is supported by the server and advertised in [CacheCapabilities.supported_compressor][build.bazel.remote.execution.v2.CacheCapabilities.supported_compressor].

uncompressed_hash and uncompressed_size refer to the [Digest][build.bazel.remote.execution.v2.Digest] of the data being uploaded, once uncompressed. Servers MUST verify that these match the uploaded data once uncompressed, and MUST return an INVALID_ARGUMENT error in the case of mismatch.

Note that when writing compressed blobs, the WriteRequest.write_offset in the initial request in a stream refers to the offset in the uncompressed form of the blob. In subsequent requests, WriteRequest.write_offset MUST be the sum of the first request’s ‘WriteRequest.write_offset’ and the total size of all the compressed data bundles in the previous requests. Note that this mixes an uncompressed offset with a compressed byte length, which is nonsensical, but it is done to fit the semantics of the existing ByteStream protocol.

Uploads of the same data MAY occur concurrently in any form, compressed or uncompressed.

Clients SHOULD NOT use gRPC-level compression for ByteStream API Write calls of compressed blobs, since this would compress already-compressed data.

When attempting an upload, if another client has already completed the upload (which may occur in the middle of a single upload if another client uploads the same blob concurrently), the request will terminate immediately without error, and with a response whose committed_size is the value -1 if this is a compressed upload, or with the full size of the uploaded file if this is an uncompressed upload (regardless of how much data was transmitted by the client). If the client completes the upload but the [Digest][build.bazel.remote.execution.v2.Digest] does not match, an INVALID_ARGUMENT error will be returned. In either case, the client should not attempt to retry the upload.

Small downloads can be grouped and requested in a batch via [BatchReadBlobs][build.bazel.remote.execution.v2.ContentAddressableStorage.BatchReadBlobs].

For large downloads, the client must use the [Read method][google.bytestream.ByteStream.Read] of the ByteStream API.

For uncompressed data, The ReadRequest.resource_name is of the following form: {instance_name}/blobs/{hash}/{size} Where instance_name, hash and size are defined as for uploads.

Data can alternatively be downloaded in compressed form, with the following ReadRequest.resource_name form: {instance_name}/compressed-blobs/{compressor}/{uncompressed_hash}/{uncompressed_size}

Where:

instance_name and compressor are defined as for uploads.

uncompressed_hash and uncompressed_size refer to the [Digest][build.bazel.remote.execution.v2.Digest] of the data being downloaded, once uncompressed. Clients MUST verify that these match the downloaded data once uncompressed, and take appropriate steps in the case of failure such as retrying a limited number of times or surfacing an error to the user.

When downloading compressed blobs:

ReadRequest.read_offset refers to the offset in the uncompressed form of the blob.

Servers MUST return INVALID_ARGUMENT if ReadRequest.read_limit is non-zero.

Servers MAY use any compression level they choose, including different levels for different blobs (e.g. choosing a level designed for maximum speed for data known to be incompressible).

Clients SHOULD NOT use gRPC-level compression, since this would compress already-compressed data.

Servers MUST be able to provide data for all recently advertised blobs in each of the compression formats that the server supports, as well as in uncompressed form.

The lifetime of entries in the CAS is implementation specific, but it SHOULD be long enough to allow for newly-added and recently looked-up entries to be used in subsequent calls (e.g. to [Execute][build.bazel.remote.execution.v2.Execution.Execute]).

Servers MUST behave as though empty blobs are always available, even if they have not been uploaded. Clients MAY optimize away the uploading or downloading of empty blobs.

As with other services in the Remote Execution API, any call may return an error with a [RetryInfo][google.rpc.RetryInfo] error detail providing information about when the client should retry the request; clients SHOULD respect the information provided.

Implementations

impl ContentAddressableStorageClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error>
where D: TryInto<Endpoint>, D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> ContentAddressableStorageClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> ContentAddressableStorageClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn find_missing_blobs( &mut self, request: impl IntoRequest<FindMissingBlobsRequest>, ) -> Result<Response<FindMissingBlobsResponse>, Status>

Determine if blobs are present in the CAS.

Clients can use this API before uploading blobs to determine which ones are already present in the CAS and do not need to be uploaded again.

Servers SHOULD increase the lifetimes of the referenced blobs if necessary and applicable.

There are no method-specific errors.

pub async fn batch_update_blobs( &mut self, request: impl IntoRequest<BatchUpdateBlobsRequest>, ) -> Result<Response<BatchUpdateBlobsResponse>, Status>

Upload many blobs at once.

The server may enforce a limit of the combined total size of blobs to be uploaded using this API. This limit may be obtained using the [Capabilities][build.bazel.remote.execution.v2.Capabilities] API. Requests exceeding the limit should either be split into smaller chunks or uploaded using the [ByteStream API][google.bytestream.ByteStream], as appropriate.

This request is equivalent to calling a Bytestream Write request on each individual blob, in parallel. The requests may succeed or fail independently.

Errors:

INVALID_ARGUMENT: The client attempted to upload more than the server supported limit.

Individual requests may return the following errors, additionally:

RESOURCE_EXHAUSTED: There is insufficient disk quota to store the blob.

INVALID_ARGUMENT: The [Digest][build.bazel.remote.execution.v2.Digest] does not match the provided data.

pub async fn batch_read_blobs( &mut self, request: impl IntoRequest<BatchReadBlobsRequest>, ) -> Result<Response<BatchReadBlobsResponse>, Status>

Download many blobs at once.

The server may enforce a limit of the combined total size of blobs to be downloaded using this API. This limit may be obtained using the [Capabilities][build.bazel.remote.execution.v2.Capabilities] API. Requests exceeding the limit should either be split into smaller chunks or downloaded using the [ByteStream API][google.bytestream.ByteStream], as appropriate.

This request is equivalent to calling a Bytestream Read request on each individual blob, in parallel. The requests may succeed or fail independently.

Errors:

INVALID_ARGUMENT: The client attempted to read more than the server supported limit.

Every error on individual read will be returned in the corresponding digest status.

pub async fn get_tree( &mut self, request: impl IntoRequest<GetTreeRequest>, ) -> Result<Response<Streaming<GetTreeResponse>>, Status>

Fetch the entire directory tree rooted at a node.

This request must be targeted at a [Directory][build.bazel.remote.execution.v2.Directory] stored in the [ContentAddressableStorage][build.bazel.remote.execution.v2.ContentAddressableStorage] (CAS). The server will enumerate the Directory tree recursively and return every node descended from the root.

The GetTreeRequest.page_token parameter can be used to skip ahead in the stream (e.g. when retrying a partially completed and aborted request), by setting it to a value taken from GetTreeResponse.next_page_token of the last successfully processed GetTreeResponse).

The exact traversal order is unspecified and, unless retrieving subsequent pages from an earlier request, is not guaranteed to be stable across multiple invocations of GetTree.

If part of the tree is missing from the CAS, the server will return the portion present and omit the rest.

Errors:

NOT_FOUND: The requested tree root is not present in the CAS.