re_grpc_proto::build::bazel::remote::execution::v2::content_addressable_storage_server

Trait ContentAddressableStorage

pub trait ContentAddressableStorage:
    Send
    + Sync
    + 'static {
    type GetTreeStream: Stream<Item = Result<GetTreeResponse, Status>> + Send + 'static;

    // Required methods
    fn find_missing_blobs<'life0, 'async_trait>(
        &'life0 self,
        request: Request<FindMissingBlobsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<FindMissingBlobsResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn batch_update_blobs<'life0, 'async_trait>(
        &'life0 self,
        request: Request<BatchUpdateBlobsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<BatchUpdateBlobsResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn batch_read_blobs<'life0, 'async_trait>(
        &'life0 self,
        request: Request<BatchReadBlobsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<BatchReadBlobsResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_tree<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetTreeRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::GetTreeStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with ContentAddressableStorageServer.

Required Associated Types

type GetTreeStream: Stream<Item = Result<GetTreeResponse, Status>> + Send + 'static

Server streaming response type for the GetTree method.

Required Methods

fn find_missing_blobs<'life0, 'async_trait>( &'life0 self, request: Request<FindMissingBlobsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<FindMissingBlobsResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Determine if blobs are present in the CAS.

Clients can use this API before uploading blobs to determine which ones are already present in the CAS and do not need to be uploaded again.

Servers SHOULD increase the lifetimes of the referenced blobs if necessary and applicable.

There are no method-specific errors.

fn batch_update_blobs<'life0, 'async_trait>( &'life0 self, request: Request<BatchUpdateBlobsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<BatchUpdateBlobsResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Upload many blobs at once.

The server may enforce a limit of the combined total size of blobs to be uploaded using this API. This limit may be obtained using the [Capabilities][build.bazel.remote.execution.v2.Capabilities] API. Requests exceeding the limit should either be split into smaller chunks or uploaded using the [ByteStream API][google.bytestream.ByteStream], as appropriate.

This request is equivalent to calling a Bytestream Write request on each individual blob, in parallel. The requests may succeed or fail independently.

Errors:

INVALID_ARGUMENT: The client attempted to upload more than the server supported limit.

Individual requests may return the following errors, additionally:

RESOURCE_EXHAUSTED: There is insufficient disk quota to store the blob.

INVALID_ARGUMENT: The [Digest][build.bazel.remote.execution.v2.Digest] does not match the provided data.

fn batch_read_blobs<'life0, 'async_trait>( &'life0 self, request: Request<BatchReadBlobsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<BatchReadBlobsResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Download many blobs at once.

The server may enforce a limit of the combined total size of blobs to be downloaded using this API. This limit may be obtained using the [Capabilities][build.bazel.remote.execution.v2.Capabilities] API. Requests exceeding the limit should either be split into smaller chunks or downloaded using the [ByteStream API][google.bytestream.ByteStream], as appropriate.

This request is equivalent to calling a Bytestream Read request on each individual blob, in parallel. The requests may succeed or fail independently.

Errors:

INVALID_ARGUMENT: The client attempted to read more than the server supported limit.

Every error on individual read will be returned in the corresponding digest status.

fn get_tree<'life0, 'async_trait>( &'life0 self, request: Request<GetTreeRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::GetTreeStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Fetch the entire directory tree rooted at a node.

This request must be targeted at a [Directory][build.bazel.remote.execution.v2.Directory] stored in the [ContentAddressableStorage][build.bazel.remote.execution.v2.ContentAddressableStorage] (CAS). The server will enumerate the Directory tree recursively and return every node descended from the root.

The GetTreeRequest.page_token parameter can be used to skip ahead in the stream (e.g. when retrying a partially completed and aborted request), by setting it to a value taken from GetTreeResponse.next_page_token of the last successfully processed GetTreeResponse).

The exact traversal order is unspecified and, unless retrieving subsequent pages from an earlier request, is not guaranteed to be stable across multiple invocations of GetTree.

If part of the tree is missing from the CAS, the server will return the portion present and omit the rest.

Errors:

NOT_FOUND: The requested tree root is not present in the CAS.