re_grpc_proto::google::bytestream::byte_stream_server

Trait ByteStream

pub trait ByteStream:
    Send
    + Sync
    + 'static {
    type ReadStream: Stream<Item = Result<ReadResponse, Status>> + Send + 'static;

    // Required methods
    fn read<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ReadRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::ReadStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn write<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Streaming<WriteRequest>>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<WriteResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn query_write_status<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryWriteStatusRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<QueryWriteStatusResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with ByteStreamServer.

Required Associated Types

type ReadStream: Stream<Item = Result<ReadResponse, Status>> + Send + 'static

Server streaming response type for the Read method.

Required Methods

fn read<'life0, 'async_trait>( &'life0 self, request: Request<ReadRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::ReadStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Read() is used to retrieve the contents of a resource as a sequence of bytes. The bytes are returned in a sequence of responses, and the responses are delivered as the results of a server-side streaming RPC.

fn write<'life0, 'async_trait>( &'life0 self, request: Request<Streaming<WriteRequest>>, ) -> Pin<Box<dyn Future<Output = Result<Response<WriteResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Write() is used to send the contents of a resource as a sequence of bytes. The bytes are sent in a sequence of request protos of a client-side streaming RPC.

A Write() action is resumable. If there is an error or the connection is broken during the Write(), the client should check the status of the Write() by calling QueryWriteStatus() and continue writing from the returned committed_size. This may be less than the amount of data the client previously sent.

Calling Write() on a resource name that was previously written and finalized could cause an error, depending on whether the underlying service allows over-writing of previously written resources.

When the client closes the request channel, the service will respond with a WriteResponse. The service will not view the resource as complete until the client has sent a WriteRequest with finish_write set to true. Sending any requests on a stream after sending a request with finish_write set to true will cause an error. The client should check the WriteResponse it receives to determine how much data the service was able to commit and whether the service views the resource as complete or not.

fn query_write_status<'life0, 'async_trait>( &'life0 self, request: Request<QueryWriteStatusRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<QueryWriteStatusResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

QueryWriteStatus() is used to find the committed_size for a resource that is being written, which can then be used as the write_offset for the next Write() call.

If the resource does not exist (i.e., the resource has been deleted, or the first Write() has not yet reached the service), this method returns the error NOT_FOUND.

The client may call QueryWriteStatus() at any time to determine how much data has been processed for this resource. This is useful if the client is buffering data and needs to know which data can be safely evicted. For any sequence of QueryWriteStatus() calls for a given resource name, the sequence of returned committed_size values will be non-decreasing.