re_grpc_proto::google::longrunning::operations_server

Trait Operations

pub trait Operations:
    Send
    + Sync
    + 'static {
    // Required methods
    fn list_operations<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ListOperationsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ListOperationsResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_operation<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetOperationRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Operation>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn delete_operation<'life0, 'async_trait>(
        &'life0 self,
        request: Request<DeleteOperationRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn cancel_operation<'life0, 'async_trait>(
        &'life0 self,
        request: Request<CancelOperationRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn wait_operation<'life0, 'async_trait>(
        &'life0 self,
        request: Request<WaitOperationRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Operation>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with OperationsServer.

Required Methods

fn list_operations<'life0, 'async_trait>( &'life0 self, request: Request<ListOperationsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ListOperationsResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Lists operations that match the specified filter in the request. If the server doesn’t support this method, it returns UNIMPLEMENTED.

NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

fn get_operation<'life0, 'async_trait>( &'life0 self, request: Request<GetOperationRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Operation>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

fn delete_operation<'life0, 'async_trait>( &'life0 self, request: Request<DeleteOperationRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn’t support this method, it returns google.rpc.Code.UNIMPLEMENTED.

fn cancel_operation<'life0, 'async_trait>( &'life0 self, request: Request<CancelOperationRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn’t support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use [Operations.GetOperation][google.longrunning.Operations.GetOperation] or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an [Operation.error][google.longrunning.Operation.error] value with a [google.rpc.Status.code][google.rpc.Status.code] of 1, corresponding to Code.CANCELLED.

fn wait_operation<'life0, 'async_trait>( &'life0 self, request: Request<WaitOperationRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Operation>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Waits until the specified long-running operation is done or reaches at most a specified timeout, returning the latest state. If the operation is already done, the latest state is immediately returned. If the timeout specified is greater than the default HTTP/RPC timeout, the HTTP/RPC timeout is used. If the server does not support this method, it returns google.rpc.Code.UNIMPLEMENTED. Note that this method is on a best-effort basis. It may return the latest state before the specified timeout (including immediately), meaning even an immediate response is no guarantee that the operation is done.